+++ arviz/plots/backends/matplotlib/pairplot.py
@@ -39,6 +39,8 @@ def plot_pair(
     point_estimate,
     point_estimate_kwargs,
     point_estimate_marker_kwargs,
+    reference_values,
+    reference_values_kwargs,
 ):
     """Matplotlib pairplot."""
     if backend_kwargs is None:
@@ -59,6 +61,35 @@ def plot_pair(
         kde_kwargs.setdefault("contourf_kwargs", {"alpha": 0})
         kde_kwargs.setdefault("contour_kwargs", {})
         kde_kwargs["contour_kwargs"].setdefault("colors", "k")
+
+    if reference_values:
+        reference_values_copy = {}
+        label = []
+        for variable in list(reference_values.keys()):
+            if " " in variable:
+                variable_copy = variable.replace(" ", "\n", 1)
+            else:
+                variable_copy = variable
+
+            label.append(variable_copy)
+            reference_values_copy[variable_copy] = reference_values[variable]
+
+        difference = set(flat_var_names).difference(set(label))
+
+        if difference:
+            warn = [dif.replace("\n", " ", 1) for dif in difference]
+            warnings.warn(
+                "Argument reference_values does not include reference value for: {}".format(
+                    ", ".join(warn)
+                ),
+                UserWarning,
+            )
+
+    if reference_values_kwargs is None:
+        reference_values_kwargs = {}
+
+    reference_values_kwargs.setdefault("color", "C3")
+    reference_values_kwargs.setdefault("marker", "o")
 
     # pylint: disable=too-many-nested-blocks
     if numvars == 2:
@@ -140,6 +171,12 @@ def plot_pair(
 
             ax.scatter(pe_x, pe_y, marker="s", s=figsize[0] + 50, **point_estimate_kwargs, zorder=4)
 
+        if reference_values:
+            ax.plot(
+                reference_values_copy[flat_var_names[0]],
+                reference_values_copy[flat_var_names[1]],
+                **reference_values_kwargs,
+            )
         ax.set_xlabel("{}".format(flat_var_names[0]), fontsize=ax_labelsize, wrap=True)
         ax.set_ylabel("{}".format(flat_var_names[1]), fontsize=ax_labelsize, wrap=True)
         ax.tick_params(labelsize=xt_labelsize)
@@ -228,6 +265,16 @@ def plot_pair(
                         ax[j, i].scatter(
                             pe_x, pe_y, s=figsize[0] + 50, zorder=4, **point_estimate_marker_kwargs
                         )
+
+                    if reference_values:
+                        x_name = flat_var_names[i]
+                        y_name = flat_var_names[j]
+                        if x_name and y_name not in difference:
+                            ax[j, i].plot(
+                                reference_values_copy[x_name],
+                                reference_values_copy[y_name],
+                                **reference_values_kwargs,
+                            )
 
                 if j != numvars - 1:
                     ax[j, i].axes.get_xaxis().set_major_formatter(NullFormatter())